[feat]: reduce fonts
